+++ client/src/pages/dish/components/createBlockPhotos.tsx
@@ -1,0 +1,31 @@
+import {useEffect, useRef, useState} from "react";
+import {Icon} from "@iconify/react";
+
+export default function CreateBlockPhotos ({dish}) {
+    const [photo, setPhoto] = useState(0)
+    const myRef = useRef(null);
+
+    useEffect(() => {
+        const container = myRef.current;
+
+        function handleWheel (e){
+            e.preventDefault()
+            container.scrollLeft += e.deltaY;
+        }
+
+        container.addEventListener("wheel", handleWheel, { passive: false });
+        return () => container.removeEventListener("wheel", handleWheel);
+    }, []);
+
+    return <div className="product__photos">
+        <div className="product__photo-spicy">
+            <img src={dish.dish_images[photo]?.image_url} alt={dish.dish_images[photo]?.title}/>
+            {dish.spicy === true && <Icon icon="mdi:fire" className="icon"/>}
+        </div>
+        <div className="product__additional-photos" ref={myRef}>
+            {dish.dish_images.map((image, i) =>
+                <img src={image.image_url} alt={image.title} key={image.id} onClick={() => setPhoto(i)} />
+            )}
+        </div>
+    </div>
+}
